Altair light client sync protocol

+++ beacon-chain/rpc/eth/beacon/light_client.go
@@ -148,7 +148,7 @@ func (bs *Server) GetLightClientFinalityUpdate(ctx context.Context, _ *empty.Emp
 		FinalizedHeader: finalizedHeader,
 		FinalityBranch:  finalityBranch,
 		SyncAggregate:   &syncAggregateV1,
-		SignatureSlot:   uint64(prevSlot),
+		SignatureSlot:   prevSlot,
 	}
 
 	// Return the result
@@ -206,7 +206,7 @@ func (bs *Server) GetLightClientOptimisticUpdate(ctx context.Context, _ *empty.E
 	data := ethpbv2.OptimisticUpdate{
 		AttestedHeader: attestedHeader,
 		SyncAggregate:  &syncAggregateV1,
-		SignatureSlot:  uint64(prevSlot),
+		SignatureSlot:  prevSlot,
 	}
 
 	// Return the result
